VMM/IEM: Nested VMX: bugref:9180 nit.
svn:sync-xref-src-repo-rev: r125594

+++ src/VBox/VMM/VMMAll/IEMAllCImplVmxInstr.cpp.h
@@ -1,4 +1,4 @@
-/* $Id: IEMAllCImplVmxInstr.cpp.h 74683 2018-10-08 15:13:52Z ramshankar.venkataraman@oracle.com $ */
+/* $Id: IEMAllCImplVmxInstr.cpp.h 74684 2018-10-08 15:15:05Z ramshankar.venkataraman@oracle.com $ */
 /** @file
  * IEM - VT-x instruction implementation.
  */
@@ -3390,7 +3390,7 @@ IEM_STATIC VBOXSTRICTRC iemVmxVmexitTaskSwitch(PVMCPU pVCpu, IEMTASKSWITCH enmTa
      *
      * See Intel spec. 25.2 "Other Causes Of VM Exits".
      */
-    uint8_t uTaskSwitchSrc;
+    uint32_t uTaskSwitchSrc;
     switch (enmTaskSwitch)
     {
         case IEMTASKSWITCH_CALL:     uTaskSwitchSrc = 0; break;
